[Stream] Add buffered tests
Create a single testing stub for running a stream in buffered mode,
and checks the final state. Used for testing:
1. Power (1, 1, 10, 1)
2. Exp(2)
3. Exp(1.5)

Currently tested for 10k steps, to check for scalability.

+++ core/db_test.go
@@ -209,9 +209,100 @@ func TestDBAppendAfterRead(t *testing.T) {
 	}
 }
 
+func testStub(t *testing.T,
+	dbPath string,
+	timesteps int64,
+	seq window.LengthsSequence,
+	expectedWindows int) {
+
+	var streamId int64
+	{
+		config := StoreConfig{
+			EachBufferSize:  32,
+			NumBuffer:       8,
+			WindowsPerMerge: 8,
+		}
+		err := os.RemoveAll(dbPath)
+		assert.NoError(t, err)
+		db, err := New(dbPath)
+		assert.NoError(t, err)
+		stream, err := db.NewStream([]string{"count", "sum", "max"}, seq)
+		assert.NoError(t, err)
+		stream.SetConfig(&config)
+		ctx, cancelFunc := context.WithCancel(context.Background())
+		// TODO: Let `db` start the stream, and keep track of the ctxs?
+		err = stream.Run(ctx)
+		assert.NoError(t, err)
+		streamId = stream.streamId
+		for i := int64(0); i < timesteps; i++ {
+			err := stream.Append(i, 2*float64(i))
+			assert.NoError(t, err)
+		}
+
+		err = stream.Flush()
+		assert.NoError(t, err)
+		err = db.Close()
+		assert.NoError(t, err)
+		cancelFunc()
+	}
+	{
+		params := QueryParams{
+			ConfidenceLevel: 0.95,
+			SDMultiplier:    1.0,
+		}
+		db, err := Open(dbPath)
+		assert.NoError(t, err)
+		stream, err := db.GetStream(streamId)
+		assert.NoError(t, err)
+		{
+			result, err := stream.Query("count", 0, timesteps-1, &params)
+			assert.NoError(t, err)
+			assert.Equal(t, result.value.Count.Value, float64(timesteps))
+			assert.Equal(t, result.error, 0.0)
+		}
+		{
+			result, err := stream.Query("sum", 0, timesteps-1, &params)
+			assert.NoError(t, err)
+			assert.Equal(t, result.value.Sum.Value, float64((timesteps-1)*timesteps))
+			assert.Equal(t, result.error, 0.0)
+		}
+		{
+			result, err := stream.Query("max", 0, timesteps-1, &params)
+			assert.NoError(t, err)
+			assert.Equal(t, result.value.Max.Value, 2*float64(timesteps-1))
+		}
+		numSummaryWindows, err := stream.manager.GetSummaryWindowInRange(0, timesteps)
+		assert.NoError(t, err)
+		assert.Equal(t, len(numSummaryWindows), expectedWindows)
+	}
+}
+
+func TestDB_Power(t *testing.T) {
+	dbPath := "testdb4"
+	seq := window.NewPowerLengthsSequence(1, 1, 10, 1)
+	testStub(t, dbPath, 10000, seq, 598)
+}
+
+func TestDB_Exp2(t *testing.T) {
+	dbPath := "testdb5"
+	seq := window.NewExponentialLengthsSequence(2)
+	testStub(t, dbPath, 10000, seq, 18)
+}
+
+func TestDB_Exp1Point5(t *testing.T) {
+	dbPath := "testdb6"
+	seq := window.NewExponentialLengthsSequence(1.5)
+	testStub(t, dbPath, 10000, seq, 33)
+}
+
 func BenchmarkDB_Append(b *testing.B) {
 	dbPath := "testdb_bm1"
 	nStreams := 8
+
+	if b.N/nStreams < 1 {
+		return
+	}
+
 	err := os.RemoveAll(dbPath)
 	if err != nil {
 		b.FailNow()
@@ -255,6 +346,11 @@ func BenchmarkDB_Append(b *testing.B) {
 func BenchmarkDB_Append_Buffered(b *testing.B) {
 	dbPath := "testdb_bm2"
 	nStreams := 64
+
+	if b.N/nStreams < 1 {
+		return
+	}
+
 	err := os.RemoveAll(dbPath)
 	if err != nil {
 		b.FailNow()
